+++ src/intrinsics.cpp
@@ -35,6 +35,16 @@ static void jl_init_intrinsic_functions_codegen(Module *m)
 #undef ALIAS
 }
 
+extern "C" JL_DLLEXPORT uint32_t jl_get_LLVM_VERSION(void);
+JL_DLLEXPORT uint32_t jl_get_LLVM_VERSION(void)
+{
+    return 10000 * LLVM_VERSION_MAJOR + 100 * LLVM_VERSION_MINOR
+#ifdef LLVM_VERSION_PATCH
+        + LLVM_VERSION_PATCH
+#endif
+        ;
+}
+
 /*
   low-level intrinsics design: TODO: fix description below
   functions like add_int expect unboxed values of matching bit-length.
@@ -512,7 +522,7 @@ static jl_cgval_t generic_unbox(jl_value_t *targ, jl_value_t *x, jl_codectx_t *c
         if (!v.ispointer)
             builder.CreateAlignedStore(emit_unbox(llvmt, v, v.typ), builder.CreatePointerCast(newobj, llvmt->getPointerTo()), alignment);
         else
-            builder.CreateMemCpy(newobj, builder.CreateBitCast(v.V, T_pint8), nb, alignment);
+            prepare_call(builder.CreateMemCpy(newobj, builder.CreateBitCast(v.V, T_pint8), nb, alignment)->getCalledValue());
         return mark_julia_type(newobj, true, bt ? bt : (jl_value_t*)jl_any_type);
     }
 
@@ -687,16 +697,13 @@ static jl_cgval_t emit_checked_fptoui(jl_value_t *targ, jl_value_t *x, jl_codect
 
 static jl_cgval_t emit_runtime_pointerref(jl_value_t *e, jl_value_t *i, jl_codectx_t *ctx)
 {
-    Value *preffunc = // TODO: move this to the codegen initialization code section
-        jl_Module->getOrInsertFunction("jl_pointerref",
-                                       FunctionType::get(T_pjlvalue, two_pvalue_llvmt, false));
     int ldepth = ctx->gc.argDepth;
     jl_cgval_t parg = emit_boxed_rooted(e, ctx);
     Value *iarg = boxed(emit_expr(i, ctx), ctx);
 #ifdef LLVM37
-    Value *ret = builder.CreateCall(prepare_call(preffunc), { parg.V, iarg });
+    Value *ret = builder.CreateCall(prepare_call(jlpref_func), { parg.V, iarg });
 #else
-    Value *ret = builder.CreateCall2(prepare_call(preffunc), parg.V, iarg);
+    Value *ret = builder.CreateCall2(prepare_call(jlpref_func), parg.V, iarg);
 #endif
     ctx->gc.argDepth = ldepth;
     jl_value_t *ety;
@@ -745,8 +752,8 @@ static jl_cgval_t emit_pointerref(jl_value_t *e, jl_value_t *i, jl_codectx_t *ct
         im1 = builder.CreateMul(im1, ConstantInt::get(T_size,
                     LLT_ALIGN(size, ((jl_datatype_t*)ety)->alignment)));
         thePtr = builder.CreateGEP(builder.CreateBitCast(thePtr, T_pint8), im1);
-        builder.CreateMemCpy(builder.CreateBitCast(strct, T_pint8),
-                             thePtr, size, 1);
+        prepare_call(builder.CreateMemCpy(builder.CreateBitCast(strct, T_pint8),
+                             thePtr, size, 1)->getCalledValue());
         return mark_julia_type(strct, true, ety);
     }
     // TODO: alignment?
@@ -755,17 +762,14 @@ static jl_cgval_t emit_pointerref(jl_value_t *e, jl_value_t *i, jl_codectx_t *ct
 
 static jl_cgval_t emit_runtime_pointerset(jl_value_t *e, jl_value_t *x, jl_value_t *i, jl_codectx_t *ctx)
 {
-    Value *psetfunc = // TODO: move this to the codegen initialization code section
-        jl_Module->getOrInsertFunction("jl_pointerset",
-                                       FunctionType::get(T_pjlvalue, three_pvalue_llvmt, false));
     int ldepth = ctx->gc.argDepth;
     jl_cgval_t parg = emit_boxed_rooted(e, ctx);
     Value *iarg = emit_boxed_rooted(i, ctx).V;
     Value *xarg = boxed(emit_expr(x, ctx), ctx);
 #ifdef LLVM37
-    builder.CreateCall(prepare_call(psetfunc), { parg.V, xarg, iarg });
+    builder.CreateCall(prepare_call(jlpset_func), { parg.V, xarg, iarg });
 #else
-    builder.CreateCall3(prepare_call(psetfunc), parg.V, xarg, iarg);
+    builder.CreateCall3(prepare_call(jlpset_func), parg.V, xarg, iarg);
 #endif
     ctx->gc.argDepth = ldepth;
     return parg;
@@ -808,8 +812,8 @@ static jl_cgval_t emit_pointerset(jl_value_t *e, jl_value_t *x, jl_value_t *i, j
         uint64_t size = ((jl_datatype_t*)ety)->size;
         im1 = builder.CreateMul(im1, ConstantInt::get(T_size,
                     LLT_ALIGN(size, ((jl_datatype_t*)ety)->alignment)));
-        builder.CreateMemCpy(builder.CreateGEP(builder.CreateBitCast(thePtr, T_pint8), im1),
-                             builder.CreateBitCast(val.V, T_pint8), size, 1);
+        prepare_call(builder.CreateMemCpy(builder.CreateGEP(builder.CreateBitCast(thePtr, T_pint8), im1),
+                             builder.CreateBitCast(val.V, T_pint8), size, 1)->getCalledValue());
     }
     else {
         if (!emitted) {
@@ -824,7 +828,7 @@ static jl_cgval_t emit_pointerset(jl_value_t *e, jl_value_t *x, jl_value_t *i, j
     return mark_julia_type(thePtr, false, aty);
 }
 
-static Value *emit_checked_srem(Value *x, Value *den, jl_codectx_t *ctx)
+static Value *emit_checked_srem_int(Value *x, Value *den, jl_codectx_t *ctx)
 {
     Type *t = den->getType();
     raise_exception_unless(builder.CreateICmpNE(den, ConstantInt::get(t,0)),
@@ -1157,12 +1161,12 @@ static Value *emit_untyped_intrinsic(intrinsic f, Value *x, Value *y, Value *z, 
     case fma_float: {
       assert(y->getType() == x->getType());
       assert(z->getType() == y->getType());
-      Value *fmaintr = Intrinsic::getDeclaration(jl_Module, Intrinsic::fma,
+      Value *fmaintr = Intrinsic::getDeclaration(builtins_module, Intrinsic::fma,
                                    ArrayRef<Type*>(x->getType()));
 #ifdef LLVM37
-      return builder.CreateCall(fmaintr,{ FP(x), FP(y), FP(z) });
+      return builder.CreateCall(prepare_call(fmaintr),{ FP(x), FP(y), FP(z) });
 #else
-      return builder.CreateCall3(fmaintr, FP(x), FP(y), FP(z));
+      return builder.CreateCall3(prepare_call(fmaintr), FP(x), FP(y), FP(z));
 #endif
     }
     case muladd_float:
@@ -1175,8 +1179,8 @@ static Value *emit_untyped_intrinsic(intrinsic f, Value *x, Value *y, Value *z, 
 #else
       return builder.CreateCall3
 #endif
-        (Intrinsic::getDeclaration(jl_Module, Intrinsic::fmuladd,
-                                   ArrayRef<Type*>(x->getType())),
+        (prepare_call(Intrinsic::getDeclaration(builtins_module, Intrinsic::fmuladd,
+                                   ArrayRef<Type*>(x->getType()))),
          #ifdef LLVM37
          {FP(x), FP(y), FP(z)}
          #else
@@ -1189,30 +1193,30 @@ static Value *emit_untyped_intrinsic(intrinsic f, Value *x, Value *y, Value *z, 
         CreateFAdd(builder.CreateFMul(FP(x), FP(y)), FP(z));
 #endif
 
-    case checked_sadd:
-    case checked_uadd:
-    case checked_ssub:
-    case checked_usub:
-    case checked_smul:
-    case checked_umul: {
+    case checked_sadd_int:
+    case checked_uadd_int:
+    case checked_ssub_int:
+    case checked_usub_int:
+    case checked_smul_int:
+    case checked_umul_int: {
         Value *ix = JL_INT(x); Value *iy = JL_INT(y);
         assert(ix->getType() == iy->getType());
         Value *intr =
-            Intrinsic::getDeclaration(jl_Module,
-               f==checked_sadd ?
-               Intrinsic::sadd_with_overflow :
-               (f==checked_uadd ?
-                Intrinsic::uadd_with_overflow :
-                (f==checked_ssub ?
-                 Intrinsic::ssub_with_overflow :
-                 (f==checked_usub ?
-                  Intrinsic::usub_with_overflow :
-                  (f==checked_smul ?
-                   Intrinsic::smul_with_overflow :
-                   Intrinsic::umul_with_overflow)))),
-               ArrayRef<Type*>(ix->getType()));
+            Intrinsic::getDeclaration(builtins_module,
+                f==checked_sadd_int ?
+                Intrinsic::sadd_with_overflow :
+                (f==checked_uadd_int ?
+                 Intrinsic::uadd_with_overflow :
+                 (f==checked_ssub_int ?
+                  Intrinsic::ssub_with_overflow :
+                  (f==checked_usub_int ?
+                   Intrinsic::usub_with_overflow :
+                   (f==checked_smul_int ?
+                    Intrinsic::smul_with_overflow :
+                    Intrinsic::umul_with_overflow)))),
+                ArrayRef<Type*>(ix->getType()));
 #ifdef LLVM37
-        Value *res = builder.CreateCall(intr,{ix, iy});
+        Value *res = builder.CreateCall(prepare_call(intr),{ix, iy});
 #else
         Value *res = builder.CreateCall2(intr, ix, iy);
 #endif
@@ -1221,7 +1225,7 @@ static Value *emit_untyped_intrinsic(intrinsic f, Value *x, Value *y, Value *z, 
         return builder.CreateExtractValue(res, ArrayRef<unsigned>(0));
     }
 
-    case checked_sdiv:
+    case checked_sdiv_int:
         den = JL_INT(y);
         t = den->getType();
         x = JL_INT(x);
@@ -1239,17 +1243,17 @@ static Value *emit_untyped_intrinsic(intrinsic f, Value *x, Value *y, Value *z, 
                                prepare_global(jldiverr_var), ctx);
 
         return builder.CreateSDiv(x, den);
-    case checked_udiv:
+    case checked_udiv_int:
         den = JL_INT(y);
         t = den->getType();
         raise_exception_unless(builder.CreateICmpNE(den, ConstantInt::get(t,0)),
                                prepare_global(jldiverr_var), ctx);
         return builder.CreateUDiv(JL_INT(x), den);
 
-    case checked_srem:
-        return emit_checked_srem(JL_INT(x), JL_INT(y), ctx);
+    case checked_srem_int:
+        return emit_checked_srem_int(JL_INT(x), JL_INT(y), ctx);
 
-    case checked_urem:
+    case checked_urem_int:
         den = JL_INT(y);
         t = den->getType();
         raise_exception_unless(builder.CreateICmpNE(den, ConstantInt::get(t,0)),
@@ -1352,21 +1356,21 @@ static Value *emit_untyped_intrinsic(intrinsic f, Value *x, Value *y, Value *z, 
                          builder.CreateAShr(x, uint_cnvt(t,y)));
     case bswap_int:
         x = JL_INT(x);
-        return builder.CreateCall(
-            Intrinsic::getDeclaration(jl_Module, Intrinsic::bswap,
-                                      ArrayRef<Type*>(x->getType())), x);
+        return builder.CreateCall(prepare_call(
+            Intrinsic::getDeclaration(builtins_module, Intrinsic::bswap,
+                                      ArrayRef<Type*>(x->getType()))), x);
     case ctpop_int:
         x = JL_INT(x);
-        return builder.CreateCall(
-            Intrinsic::getDeclaration(jl_Module, Intrinsic::ctpop,
-                                      ArrayRef<Type*>(x->getType())), x);
+        return builder.CreateCall(prepare_call(
+            Intrinsic::getDeclaration(builtins_module, Intrinsic::ctpop,
+                                      ArrayRef<Type*>(x->getType()))), x);
     case ctlz_int: {
         x = JL_INT(x);
         Type *types[1] = {x->getType()};
-        Value *ctlz = Intrinsic::getDeclaration(jl_Module, Intrinsic::ctlz,
+        Value *ctlz = Intrinsic::getDeclaration(builtins_module, Intrinsic::ctlz,
                                       ArrayRef<Type*>(types));
 #ifdef LLVM37
-        return builder.CreateCall(ctlz, {x, ConstantInt::get(T_int1,0)});
+        return builder.CreateCall(prepare_call(ctlz), {x, ConstantInt::get(T_int1,0)});
 #else
         return builder.CreateCall2(ctlz, x, ConstantInt::get(T_int1,0));
 #endif
@@ -1374,9 +1378,9 @@ static Value *emit_untyped_intrinsic(intrinsic f, Value *x, Value *y, Value *z, 
     case cttz_int: {
         x = JL_INT(x);
         Type *types[1] = {x->getType()};
-        Value *cttz = Intrinsic::getDeclaration(jl_Module, Intrinsic::cttz, ArrayRef<Type*>(types));
+        Value *cttz = Intrinsic::getDeclaration(builtins_module, Intrinsic::cttz, ArrayRef<Type*>(types));
 #ifdef LLVM37
-        return builder.CreateCall(cttz, {x, ConstantInt::get(T_int1, 0)});
+        return builder.CreateCall(prepare_call(cttz), {x, ConstantInt::get(T_int1, 0)});
 #else
         return builder.CreateCall2(cttz, x, ConstantInt::get(T_int1, 0));
 #endif
@@ -1395,8 +1399,9 @@ static Value *emit_untyped_intrinsic(intrinsic f, Value *x, Value *y, Value *z, 
     {
         x = FP(x);
 #ifdef LLVM34
-        return builder.CreateCall(Intrinsic::getDeclaration(jl_Module, Intrinsic::fabs,
-                                                            ArrayRef<Type*>(x->getType())),
+        return builder.CreateCall(prepare_call(
+            Intrinsic::getDeclaration(builtins_module, Intrinsic::fabs,
+                                                            ArrayRef<Type*>(x->getType()))),
                                   x);
 #else
         Type *intt = JL_INTT(x->getType());
@@ -1447,34 +1452,34 @@ static Value *emit_untyped_intrinsic(intrinsic f, Value *x, Value *y, Value *z, 
     }
     case ceil_llvm: {
         x = FP(x);
-        return builder.CreateCall(Intrinsic::getDeclaration(jl_Module, Intrinsic::ceil,
-                                                            ArrayRef<Type*>(x->getType())),
+        return builder.CreateCall(prepare_call(Intrinsic::getDeclaration(builtins_module, Intrinsic::ceil,
+                                                            ArrayRef<Type*>(x->getType()))),
                                   x);
     }
     case floor_llvm: {
         x = FP(x);
-        return builder.CreateCall(Intrinsic::getDeclaration(jl_Module, Intrinsic::floor,
-                                                            ArrayRef<Type*>(x->getType())),
+        return builder.CreateCall(prepare_call(Intrinsic::getDeclaration(builtins_module, Intrinsic::floor,
+                                                            ArrayRef<Type*>(x->getType()))),
                                   x);
     }
     case trunc_llvm: {
         x = FP(x);
-        return builder.CreateCall(Intrinsic::getDeclaration(jl_Module, Intrinsic::trunc,
-                                                            ArrayRef<Type*>(x->getType())),
+        return builder.CreateCall(prepare_call(Intrinsic::getDeclaration(builtins_module, Intrinsic::trunc,
+                                                            ArrayRef<Type*>(x->getType()))),
                                   x);
     }
     case rint_llvm: {
         x = FP(x);
-        return builder.CreateCall(Intrinsic::getDeclaration(jl_Module, Intrinsic::rint,
-                                                            ArrayRef<Type*>(x->getType())),
+        return builder.CreateCall(prepare_call(Intrinsic::getDeclaration(builtins_module, Intrinsic::rint,
+                                                            ArrayRef<Type*>(x->getType()))),
                                   x);
     }
     case sqrt_llvm: {
         x = FP(x);
         raise_exception_unless(builder.CreateFCmpUGE(x, ConstantFP::get(x->getType(),0.0)),
                                prepare_global(jldomerr_var), ctx);
-        return builder.CreateCall(Intrinsic::getDeclaration(jl_Module, Intrinsic::sqrt,
-                                                            ArrayRef<Type*>(x->getType())),
+        return builder.CreateCall(prepare_call(Intrinsic::getDeclaration(builtins_module, Intrinsic::sqrt,
+                                                            ArrayRef<Type*>(x->getType()))),
                                   x);
     }
     case powi_llvm: {
@@ -1483,12 +1488,12 @@ static Value *emit_untyped_intrinsic(intrinsic f, Value *x, Value *y, Value *z, 
         Type *tx = x->getType(); // TODO: LLVM expects this to be i32
 #ifdef LLVM36
         Type *ts[1] = { tx };
-        Value *powi = Intrinsic::getDeclaration(jl_Module, Intrinsic::powi,
+        Value *powi = Intrinsic::getDeclaration(builtins_module, Intrinsic::powi,
             ArrayRef<Type*>(ts));
 #ifdef LLVM37
-        return builder.CreateCall(powi, {x, y});
+        return builder.CreateCall(prepare_call(powi), {x, y});
 #else
-        return builder.CreateCall2(powi, x, y);
+        return builder.CreateCall2(prepare_call(powi), x, y);
 #endif
 #else
         // issue #6506
@@ -1498,8 +1503,8 @@ static Value *emit_untyped_intrinsic(intrinsic f, Value *x, Value *y, Value *z, 
     }
     case sqrt_llvm_fast: {
         x = FP(x);
-        return builder.CreateCall(Intrinsic::getDeclaration(jl_Module, Intrinsic::sqrt,
-                                                            ArrayRef<Type*>(x->getType())),
+        return builder.CreateCall(prepare_call(Intrinsic::getDeclaration(builtins_module, Intrinsic::sqrt,
+                                                            ArrayRef<Type*>(x->getType()))),
                                   x);
     }